requirements.txt
# Auto Nmap v3.0 - Python Dependencies
# =====================================
#
# REQUIRED:
#   python-nmap - Python wrapper for nmap
#
# Install with: pip3 install -r requirements.txt
#
# For Kali Linux / Debian 12+:
#   pip3 install --break-system-packages python-nmap
#   OR
#   sudo apt install python3-nmap
#
# NOTE: prettytable and colorama are NO LONGER REQUIRED
# The script now uses built-in alternatives for colors and table formatting.

# Core scanning library - Python wrapper for nmap (REQUIRED)
python-nmap>=0.7.1

# Optional: uncomment if you want enhanced table formatting
# prettytable>=3.9.0

# Optional: uncomment if native ANSI colors don't work on your system
# colorama>=0.4.6
